clean up and add colibri configs
