+++ src/services/provider/getRpcProvider.ts
@@ -2,7 +2,7 @@ import { JsonRpcApiProviderOptions, JsonRpcProvider, Network } from 'ethers'
 
 import { Network as NetworkInterface } from '../../interfaces/network'
 import getRootDomain from '../../utils/getRootDomain'
-import { VerifiedJsonRpcProvider } from './VerifiedJsonRpcProvider'
+import { ColibriRpcProvider, ColibriRpcProviderOptions, isColibriEnabledForChain } from './ColibriRpcProvider'
 
 const RPC_BATCH_CONFIG: Record<string, number> = {
   'drpc.org': 3, // batch of more than 3 requests are not allowed on free tier (response 500 with internal code 31)
@@ -22,18 +22,11 @@ const getBatchCountFromUrl = (rpcUrl: string): number | undefined => {
   }
 }
 
-const isColibriEnabledForChain = (chainId?: bigint | number) => {
-  if (!chainId) return false
-  if (process.env.USE_COLIBRI !== 'true') return false
-  // First iteration: Sepolia only
-  return Number(chainId) === 11155111
-}
-
 const getRpcProvider = (
   rpcUrls: NetworkInterface['rpcUrls'],
   chainId?: bigint | number,
   selectedRpcUrl?: string,
-  options?: JsonRpcApiProviderOptions
+  options?: ColibriRpcProviderOptions
 ) => {
   if (!rpcUrls.length) {
     throw new Error('rpcUrls must be a non-empty array')
@@ -57,7 +50,7 @@ const getRpcProvider = (
     const staticNetwork = Network.from(Number(chainId))
 
     if (isColibriEnabledForChain(chainId)) {
-      return new VerifiedJsonRpcProvider(rpcUrl, chainId, providerOptions)
+      return new ColibriRpcProvider(rpcUrl, chainId, providerOptions)
     }
 
     if (staticNetwork) {
