+++ src/services/provider/index.ts
@@ -1,1 +1,2 @@
 export * from './getRpcProvider'
+export * from './ColibriRpcProvider'
